requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.1.1
supabase==1.0.3
python-multipart==0.0.6
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1 